Code Refactoring
[*] initializing controllers with DB object,
instead of calling connecting in models

[*] Using graceful shutdown

[*] Code refactoring for code readability
