+++ src/database/pet.go
@@ -7,12 +7,6 @@ import (
 
 // GetPets returns all pets from the DB
 func (da *dataAccessor) GetPets() ([]*api.Pet, *api.Error) {
-	err := da.connect()
-	if err != nil {
-		msg := err.Error()
-		return nil, &api.Error{Message: msg, Code: 500}
-	}
-
 	var pets []models.Pet
 	var response []*api.Pet
 
@@ -30,12 +24,6 @@ func (da *dataAccessor) GetPets() ([]*api.Pet, *api.Error) {
 
 // AddPet adds a new pet to the DB
 func (da *dataAccessor) AddPet(pet *models.Pet) (*api.Pet, *api.Error) {
-	err := da.connect()
-	if err != nil {
-		msg := err.Error()
-		return nil, &api.Error{Message: msg, Code: 500}
-	}
-
 	if err := da.db.Create(&pet).Error; err != nil {
 		msg := "Failed to create pet for " + pet.Name
 		return nil, &api.Error{Message: msg, Code: 500}
@@ -45,12 +33,6 @@ func (da *dataAccessor) AddPet(pet *models.Pet) (*api.Pet, *api.Error) {
 
 // DeletePet deletes a pet from the DB
 func (da *dataAccessor) DeletePet(petId int64) *api.Error {
-	err := da.connect()
-	if err != nil {
-		msg := err.Error()
-		return &api.Error{Message: msg, Code: 500}
-	}
-
 	petDTO := &models.Pet{}
 	petDTO.ID = uint(petId)
 
@@ -64,12 +46,6 @@ func (da *dataAccessor) DeletePet(petId int64) *api.Error {
 
 // GetPetById gets a pet by id from the DB
 func (da *dataAccessor) GetPetById(petId int64) (*api.Pet, *api.Error) {
-	err := da.connect()
-	if err != nil {
-		msg := err.Error()
-		return nil, &api.Error{Message: msg, Code: 500}
-	}
-
 	var pet models.Pet
 
 	if err := da.db.Find(&pet, petId).Error; err != nil {
@@ -86,12 +62,6 @@ func (da *dataAccessor) GetPetById(petId int64) (*api.Pet, *api.Error) {
 
 // UpdatePetById updates pet in the DB
 func (da *dataAccessor) UpdatePet(petId int64, pet *models.Pet) *api.Error {
-	err := da.connect()
-	if err != nil {
-		msg := err.Error()
-		return &api.Error{Message: msg, Code: 500}
-	}
-
 	petDTO := &models.Pet{}
 	petDTO.ID = uint(petId)
